Update README.md
Canon printer is one of the best printers among many users due to its smooth functioning. Still, many a time, the problem occurs when the Canon mx490 series printer offline and you cannot use it, as it encounters the issue while connecting the canon mx490 printer to WiFi. To get a fix for the Canon Pixma mx490 wireless inkjet printer, the users must open the Control Panel and Navigate to Devices and Printers. Next, Right-click on the icon for your Canon mx490 printer and Choose the option– Set as default printer. 

Read More - https://printeranswers.com/canon-pixma-mx490-wireless-setup/